Setup build structure for typescript
