+++ package.json
@@ -4,10 +4,11 @@
   "description": "A javascript client for the Accentor API",
   "main": "dist/index.cjs.js",
   "module": "dist/index.esm.js",
+  "types": "dist/index.d.ts",
   "repository": "github:accentor/api-client-js",
   "scripts": {
     "lint": "eslint . --ext .ts --fix",
-    "build": "node ./esbuild.js"
+    "build": "node ./build.js"
   },
   "author": "Team Accentor <team@accentor.tech>",
   "contributors": [
@@ -25,6 +26,8 @@
     "eslint": "^7.30.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^3.4.0",
+    "estrella": "^1.4.1",
+    "npm-dts": "^1.3.10",
     "prettier": "^2.3.2",
     "typescript": "^4.3.5"
   },
